+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     Fixed:
     
         rna_tools_lib.py (RNAStructure) for remove_hydrogen, correct atom name position now
-    	rna_pdb_toolsx.py --get-seq works with protein fix #104	
+        rna_pdb_toolsx.py --get-seq works with protein fix #104	
 	
     Added: