+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - --bases-only and --no-backbone
     
         rna_tools_lib.py: add load_rnas to load structural files (via glob)
+
+        PyMOL4RNA.py: rmsdx (get all vs all rmsd matrix)
         
     Changed: