+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [pre-release]
+
+	Add:
+		PyMOL4RNA.py: seq * and seqsel (redundant function)
+
 # 3.7
 
     Fixed: