+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 3.8  
+
+mq:
+- review farfar2 wrapper #118
+- rna_mq_collect.py & rna_mq_predict.py
+	 
+md:
+- add --mdr #117
+
+wx:  fix pefx.py
+
 # 3.7.4 210727
 
 	Add: